requirements.txt
Pillow
arcade==3.2.0
pypresence
